feat: Add API route for user index
This commit adds a new API route for retrieving user data. The route is defined in the `api.php` file and maps to the `index` method of the `UserController` class. This allows users to fetch user data by sending a POST request to the `/user` endpoint.

The `index` method retrieves the user data based on the provided email in the request payload. If the user is found, a success response is returned with the user data. If the user is not found, an error response is returned.

This change improves the functionality of the application by providing an API endpoint for retrieving user data.